pyproject.toml
[project]
name = "luo-xiaohei-subs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.77.0",
    "imagehash>=4.3.2",
    "opencv-python>=4.13.0.90",
    "paddleocr>=3.4.0",
    "paddlepaddle>=3.3.0",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
]
